Build both static and dynamic versions of protobuf-lite
Currently we only build the static version of protobuf, even when given Dynamic to build_environment.bat. This is problematic, because the Mixxx scons script expects the dynamic version of protobuf when static_libs is false and therefore adds the PROTOBUF_USE_DLLS preprocessor definition. This causes a bunch of unresolved external symbol "__declspec(dllimport) ... " errors when building Mixxx. Fix this by building the dynamic version of protobuf when the Dynamic flag is given.