Build libsndfile projects in debug mode when solution conf is *-Debug
Currently the libsndfile solution fails to pass the configuration to its projects so that its projects are always build as if the configuration was "Static-Release". This causes the build_libsndfile script not to correctly find and copy the lib files when given the Debug flag, because it will look for the libs in a non-existing "%PLATFORM%/Static-Debug" directory.